[ci]: passing the artifacts to the analyse stage
The coverage report makes reference to the code located inside the
venv made by Tox. When the analyse stage is executed, that venv
is gone, so the coverage report points to invalid paths.
To tackle that issue, a cache is used. That means, only a job from
the test stage should fill the cache, because it would collide the
content of both in other case.

[ICSHWI-6943]