Add Moxa as supported interface type
Implemented as a child class of NATMCHTelnet.

Overrides the '_send_command' method, as the Moxa has issues
receiving commands that are more than 3 characters long.
Commands are slit into chunks of 3 characters or less, and
sent serially with a pause in between to allow the Moxa
telnet interface to process them.

[ICSHWI-7097]